celery/celery_exporter/requirements.txt
prometheus-client
PyYAML
requests